% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anom_theme}
\alias{anom_theme}
\title{The standard theme for our ANOM plot}
\usage{
anom_theme(...)
}
\arguments{
\item{...}{Additional arguments passed to anom_theme}
}
\value{
This function returns the new theme
}
\description{
This function creates a ggplot2 theme based on theme_bw()
}
