% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic.binomRDci}
\alias{anomic.binomRDci}
\title{ANOM for MCPAN: Binomial Distributions}
\usage{
\method{anomic}{binomRDci}(
  mc,
  xlabel = NULL,
  ylabel = NULL,
  printn = TRUE,
  printp = TRUE,
  pbin = NULL,
  drawplot = TRUE,
  ...
)
}
\arguments{
\item{mc}{An object of class glht}

\item{xlabel}{An optional character string specifying the label of the horizontal axis.}

\item{ylabel}{An optional character string specifying the label of the vertical axis.}

\item{printn}{A logical. Should per-group sample sizes be included in the chart? Default TRUE.}

\item{printp}{A logical. Should simultaneous p-values be included in the chart? Default TRUE.}

\item{pbin}{An object of class binomRDtest. Required if printp=TRUE. Default NULL.}

\item{drawplot}{A logical. Whether to explictly print() the resulting plot. Default is TRUE.}

\item{...}{Additional arguments passed to anomic.binomRDci}
}
\value{
This function returns a ggplot2 ANOM object
}
\description{
This function works with MCPAN. The binomRDci function in MCPAN addresses binomial distributions,
allowing calculation of simultaneous asymptotic confidence intervals. Note that
MCPAN uses normal approximation.
}
