% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic.glht}
\alias{anomic.glht}
\title{ANOM for General Linear Hypotheses}
\usage{
\method{anomic}{glht}(
  mc,
  xlabel = NULL,
  ylabel = NULL,
  printn = TRUE,
  printp = TRUE,
  drawplot = TRUE,
  epsilon = 0.001,
  ...
)
}
\arguments{
\item{mc}{An object of class glht}

\item{xlabel}{An optional character string specifying the label of the horizontal axis.}

\item{ylabel}{An optional character string specifying the label of the vertical axis.}

\item{printn}{A logical. Should per-group sample sizes be included in the chart? Default TRUE.}

\item{printp}{A logical. Should simultaneous p-values be included in the chart? Default TRUE.}

\item{drawplot}{A logical. Whether to explictly print() the resulting plot. Default is TRUE.}

\item{epsilon}{A number. Precision used in calculating displayed p-values for glht model. Default 1e-3.}

\item{...}{Additional arguments passed to anomic.glht}
}
\value{
This function returns a ggplot2 ANOM object
}
\description{
This function uses the multcomp package. A glht object supports a multitude of models
that include parametric models, glms, linear mixed effects and survival models.
At the start, we test for a 'grandMean' because this is required for the center line
of the plot and the calculations.
}
