% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic.mctp}
\alias{anomic.mctp}
\title{ANOM for mctp: Non-parametric Confidence Intervals}
\usage{
\method{anomic}{mctp}(
  mc,
  xlabel = NULL,
  ylabel = NULL,
  printn = TRUE,
  printp = TRUE,
  drawplot = TRUE,
  ...
)
}
\arguments{
\item{mc}{An object of class mctp}

\item{xlabel}{An optional character string specifying the label of the horizontal axis.}

\item{ylabel}{An optional character string specifying the label of the vertical axis.}

\item{printn}{A logical. Should per-group sample sizes be included in the chart? Default TRUE.}

\item{printp}{A logical. Should simultaneous p-values be included in the chart? Default TRUE.}

\item{drawplot}{A logical. Whether to explictly print() the resulting plot. Default is TRUE.}

\item{...}{Additional arguments passed to anomic.mctp}
}
\value{
This function returns a ggplot2 ANOM object
}
\description{
This function provides ANOM for the nparcomp package. The mctp function is used
to determine non-parametric confidence intervals, which assumes an unbalanced one-way layout.
}
