% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic}
\alias{anomic}
\title{ANOM --- Analysis of Means}
\usage{
anomic(mc, ...)
}
\arguments{
\item{mc}{An object of class glht, SimCi, mctp or binomRDci}

\item{...}{Additional arguments passed to the various methods}
}
\value{
This function returns a ggplot2 ANOM object
}
\description{
This function is a wrapper for the various supported classes.
}
