% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic.SimCi}
\alias{anomic.SimCi}
\title{ANOM for SimComp: Simultaneous Testing with Confidence Intervals.}
\usage{
\method{anomic}{SimCi}(
  mc,
  xlabel = NULL,
  ylabel = NULL,
  printn = TRUE,
  printp = TRUE,
  stdep = NULL,
  stind = NULL,
  pst = NULL,
  drawplot = TRUE,
  ...
)
}
\arguments{
\item{mc}{An object of class SimCi}

\item{xlabel}{An optional character string specifying the label of the horizontal axis.}

\item{ylabel}{An optional character string specifying the label of the vertical axis.}

\item{printn}{A logical. Should per-group sample sizes be included in the chart? Default TRUE.}

\item{printp}{A logical. Should simultaneous p-values be included in the chart? Default TRUE.}

\item{stdep}{A numerical vector giving the values of the dependent variable.}

\item{stind}{A factor specifying the levels of the independent variable.}

\item{pst}{An object of class SimTest. Only required simultaneous p-values are to be printed.}

\item{drawplot}{A logical. Whether to explictly print() the resulting plot. Default is TRUE.}

\item{...}{Additional arguments passed to anomic.SimCi}
}
\value{
This function returns a ggplot2 ANOM object
}
\description{
This function works with the SimComp package, which allows for simultaneous testing
with confidence intervals. This is only for one way experimental designs with
normally distributed primary response variables.
}
