% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic_gen}
\alias{anomic_gen}
\title{General ANOM plotting function}
\usage{
anomic_gen(
  mu,
  n = NULL,
  gm = NULL,
  lo,
  up,
  names = NULL,
  alternative = "two.sided",
  xlabel = "Group",
  ylabel = "Endpoint",
  printn = TRUE,
  p = NULL,
  whichone = NULL,
  drawplot = TRUE,
  psize = 5,
  title = NULL,
  dlstyle = "dashed",
  ...
)
}
\arguments{
\item{mu}{A numeric vector of group means}

\item{n}{A numeric vector of sample sizes per group. Either n or gm must be provided.}

\item{gm}{A single numeric value giving the grand mean of all groups. Either n or gm must be provided.}

\item{lo}{A numeric vector of lower (simultaneous) confidence interval bounds for comparisons to the grand mean.}

\item{up}{A numeric vector of upper (simultaneous) confidence interval bounds for comparisons to the grand mean.}

\item{names}{An (optional) vector of characters specifying the groups' names.}

\item{alternative}{A character string indicating the direction of the alternative hypothesis. Default is "two.sided", but may be changed to one-sided alternatives (either "greater" or "less").}

\item{xlabel}{An optional character string specifying the label of the horizontal axis.}

\item{ylabel}{An optional character string specifying the label of the vertical axis.}

\item{printn}{A logical. Should per-group sample sizes be included in the chart? Default TRUE.}

\item{p}{An (optional) numeric vector of (simultaneous) p-values to be printed.}

\item{whichone}{A string, one of "glm" or "ratio"; for glm in glht or ratios with SimCi.}

\item{drawplot}{A logical. If TRUE (default) draw the plot; otherwise don't.}

\item{psize}{A single numeric value giving the size of points}

\item{title}{A string. What to title the plot}

\item{dlstyle}{An object. Either a character string which is used as a linetype or a list of attributes to pass to geom_line}

\item{...}{Additional arguments passed to anom_gen}
}
\value{
This function returns a ggplot2 ANOM graphical object
}
\description{
This function creates a plotting object using the supplied parameters, and then
invokes ggplot()
}
