% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic_good_args}
\alias{anomic_good_args}
\title{Given a specific geometry function, list all valid arguments.}
\usage{
anomic_good_args(fun_name)
}
\arguments{
\item{fun_name}{A string that defines the name of a first-order function like "geom_rect"}
}
\value{
A list of strings representing valid arguments
}
\description{
This combines regular arguments and aesthetics into one list
}
