% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomic.R
\name{anomic_nicely}
\alias{anomic_nicely}
\title{Filter out required plot parameters from submitted list(...)}
\usage{
anomic_nicely(fun, arglist = NULL, ellipsis = NULL, exclude = c())
}
\arguments{
\item{fun}{A first-order function like geom_rect()}

\item{arglist}{Fixed arguments for fun}

\item{ellipsis}{Supplied list() of ellipsis, previously allocated using list(...)}

\item{exclude}{Parameters to exclude from the ellipsis list}
}
\value{
This function returns the output from invocation of the first, 'fun' argument
}
\description{
Suppress "Ignoring unknown parameters" by only submitting parameters relevant to supplied 'fun'.
Then invoke fun on the combination of the supplied arguments and the filtered ellipsis.
The ellipsis can contain both aesthetics and formal arguments.
}
